+++ java/src/test/java/io/cucumber/compatibilitykit/MessageOrdererTest.java
@@ -1,8 +1,12 @@
 package io.cucumber.compatibilitykit;
 
 import io.cucumber.messages.types.Envelope;
+import io.cucumber.messages.types.TestRunFinished;
+import io.cucumber.messages.types.TestRunStarted;
+import io.cucumber.messages.types.Timestamp;
 import org.junit.jupiter.api.Test;
 
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Random;
@@ -20,6 +24,16 @@ class MessageOrdererTest {
     @Test
     void worksOnEmptyLists(){
         List<Envelope> t = Collections.emptyList();
+        assertDoesNotThrow(() -> originalOrder.accept(t));
+        assertDoesNotThrow(() -> simulateParallelExecution.accept(t));
+    }
+
+    @Test
+    void worksOnListsWithoutTestRunEvents(){
+        List<Envelope> t = Arrays.asList(
+                Envelope.of(new TestRunStarted(new Timestamp(0L, 0L), "")),
+                Envelope.of(new TestRunFinished(null, false, new Timestamp(0L, 0L), null, ""))
+        );
         assertDoesNotThrow(() -> originalOrder.accept(t));
         assertDoesNotThrow(() -> simulateParallelExecution.accept(t));
     }